db_sqlx_maria::dev

Trait MCDatabase

pub trait MCDatabase:
    Send
    + Sync
    + CloneSPDatabase {

    // Required methods
    fn ping<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn register<'life0, 'life1, 'async_trait>(
        &'life0 self,
        p: &'life1 Register<'_>,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn delete_user<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn username_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn get_email<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn email_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        email: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn update_email<'life0, 'life1, 'async_trait>(
        &'life0 self,
        p: &'life1 UpdateEmail<'_>,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn get_password<'life0, 'life1, 'async_trait>(
        &'life0 self,
        l: &'life1 Login<'_>,
    ) -> Pin<Box<dyn Future<Output = Result<NameHash, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn update_password<'life0, 'life1, 'async_trait>(
        &'life0 self,
        p: &'life1 NameHash,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn update_username<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        current: &'life1 str,
        new: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn get_secret<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<Secret, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn get_secret_from_captcha<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<Secret, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn update_secret<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        secret: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn create_captcha<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        p: &'life2 CreateCaptcha<'_>,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn get_captcha_config<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<Captcha, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn get_all_user_captchas<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Captcha>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn update_captcha_metadata<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        p: &'life2 CreateCaptcha<'_>,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn update_captcha_key<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        old_key: &'life2 str,
        new_key: &'life3 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             Self: 'async_trait;
    fn add_captcha_levels<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        captcha_key: &'life2 str,
        levels: &'life3 [Level],
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             Self: 'async_trait;
    fn captcha_exists<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: Option<&'life1 str>,
        captcha_key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn delete_captcha_levels<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        captcha_key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn delete_captcha<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        captcha_key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn get_captcha_levels<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: Option<&'life1 str>,
        captcha_key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Level>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn get_captcha_cooldown<'life0, 'life1, 'async_trait>(
        &'life0 self,
        captcha_key: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<i32, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn add_traffic_pattern<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        captcha_key: &'life2 str,
        pattern: &'life3 TrafficPattern,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             Self: 'async_trait;
    fn get_traffic_pattern<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        captcha_key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<TrafficPattern, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn get_all_easy_captchas<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
        offset: usize,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<EasyCaptcha>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn delete_traffic_pattern<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        captcha_key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn create_notification<'life0, 'life1, 'async_trait>(
        &'life0 self,
        p: &'life1 AddNotification<'_>,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn get_all_unread_notifications<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Notification>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn mark_notification_read<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
        id: i32,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn record_fetch<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn record_solve<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn record_confirm<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn fetch_config_fetched<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        user: &'life1 str,
        key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<i64>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn fetch_solve<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        user: &'life1 str,
        key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<i64>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn fetch_confirm<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        user: &'life1 str,
        key: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<i64>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn analysis_save<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        captcha_id: &'life1 str,
        d: &'life2 CreatePerformanceAnalytics,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             Self: 'async_trait;
    fn analytics_fetch<'life0, 'life1, 'async_trait>(
        &'life0 self,
        captcha_id: &'life1 str,
        limit: usize,
        offset: usize,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PerformanceAnalytics>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn analytics_create_psuedo_id_if_not_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        captcha_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn analytics_get_psuedo_id_from_capmaign_id<'life0, 'life1, 'async_trait>(
        &'life0 self,
        captcha_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<String, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn analytics_get_capmaign_id_from_psuedo_id<'life0, 'life1, 'async_trait>(
        &'life0 self,
        psuedo_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<String, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn analytics_delete_all_records_for_campaign<'life0, 'life1, 'async_trait>(
        &'life0 self,
        campaign_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn analytics_get_all_psuedo_ids<'life0, 'async_trait>(
        &'life0 self,
        page: usize,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn update_max_nonce_for_level<'life0, 'life1, 'async_trait>(
        &'life0 self,
        captcha_key: &'life1 str,
        difficulty_factor: u32,
        latest_nonce: u32,
    ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn get_max_nonce_for_level<'life0, 'life1, 'async_trait>(
        &'life0 self,
        captcha_key: &'life1 str,
        difficulty_factor: u32,
    ) -> Pin<Box<dyn Future<Output = Result<u32, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait;
    fn stats_get_num_logs_under_time<'life0, 'async_trait>(
        &'life0 self,
        duration: u32,
    ) -> Pin<Box<dyn Future<Output = Result<usize, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn stats_get_entry_at_location_for_time_limit_asc<'life0, 'async_trait>(
        &'life0 self,
        duration: u32,
        location: u32,
    ) -> Pin<Box<dyn Future<Output = Result<Option<usize>, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;

    // Provided method
    fn analytics_captcha_is_published<'life0, 'life1, 'async_trait>(
        &'life0 self,
        campaign_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }

}

mCaptcha’s database requirements. To implement support for $Database, kindly implement this trait.

Required Methods

fn ping<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

ping DB

fn register<'life0, 'life1, 'async_trait>( &'life0 self, p: &'life1 Register<'_>, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

register a new user

fn delete_user<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

delete a user

fn username_exists<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

check if username exists

fn get_email<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<Option<String>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

get user email

fn email_exists<'life0, 'life1, 'async_trait>( &'life0 self, email: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

check if email exists

fn update_email<'life0, 'life1, 'async_trait>( &'life0 self, p: &'life1 UpdateEmail<'_>, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

update a user’s email

fn get_password<'life0, 'life1, 'async_trait>( &'life0 self, l: &'life1 Login<'_>, ) -> Pin<Box<dyn Future<Output = Result<NameHash, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

get a user’s password

fn update_password<'life0, 'life1, 'async_trait>( &'life0 self, p: &'life1 NameHash, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

update user’s password

fn update_username<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, current: &'life1 str, new: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

update username

fn get_secret<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<Secret, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

get a user’s secret

fn get_secret_from_captcha<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<Secret, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

get a user’s secret from a captcha key

fn update_secret<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, secret: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

update a user’s secret

fn create_captcha<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, p: &'life2 CreateCaptcha<'_>, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

create new captcha

fn get_captcha_config<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<Captcha, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Get captcha config

fn get_all_user_captchas<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<Vec<Captcha>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Get all captchas belonging to user

fn update_captcha_metadata<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, p: &'life2 CreateCaptcha<'_>, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

update captcha metadata; doesn’t change captcha key

fn update_captcha_key<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, username: &'life1 str, old_key: &'life2 str, new_key: &'life3 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, Self: 'async_trait,

update captcha key; doesn’t change metadata

fn add_captcha_levels<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, username: &'life1 str, captcha_key: &'life2 str, levels: &'life3 [Level], ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, Self: 'async_trait,

Add levels to captcha

fn captcha_exists<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: Option<&'life1 str>, captcha_key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

check if captcha exists

fn delete_captcha_levels<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, captcha_key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Delete all levels of a captcha

fn delete_captcha<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, captcha_key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Delete captcha

fn get_captcha_levels<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: Option<&'life1 str>, captcha_key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<Vec<Level>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Get captcha levels

fn get_captcha_cooldown<'life0, 'life1, 'async_trait>( &'life0 self, captcha_key: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<i32, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Get captcha’s cooldown period

fn add_traffic_pattern<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, username: &'life1 str, captcha_key: &'life2 str, pattern: &'life3 TrafficPattern, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, Self: 'async_trait,

Add traffic configuration

fn get_traffic_pattern<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, captcha_key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<TrafficPattern, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Get traffic configuration

fn get_all_easy_captchas<'life0, 'async_trait>( &'life0 self, limit: usize, offset: usize, ) -> Pin<Box<dyn Future<Output = Result<Vec<EasyCaptcha>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Get all easy captcha configurations on instance

fn delete_traffic_pattern<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, username: &'life1 str, captcha_key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

Delete traffic configuration

fn create_notification<'life0, 'life1, 'async_trait>( &'life0 self, p: &'life1 AddNotification<'_>, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

create new notification

fn get_all_unread_notifications<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<Vec<Notification>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

get all unread notifications

fn mark_notification_read<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, id: i32, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

mark a notification read

fn record_fetch<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

record PoWConfig fetches

fn record_solve<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

record PoWConfig solves

fn record_confirm<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

record PoWConfig confirms

fn fetch_config_fetched<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, user: &'life1 str, key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<Vec<i64>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

fetch PoWConfig fetches

fn fetch_solve<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, user: &'life1 str, key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<Vec<i64>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

fetch PoWConfig solves

fn fetch_confirm<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, user: &'life1 str, key: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<Vec<i64>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

fetch PoWConfig confirms

fn analysis_save<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, captcha_id: &'life1 str, d: &'life2 CreatePerformanceAnalytics, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, Self: 'async_trait,

record PoW timing

fn analytics_fetch<'life0, 'life1, 'async_trait>( &'life0 self, captcha_id: &'life1 str, limit: usize, offset: usize, ) -> Pin<Box<dyn Future<Output = Result<Vec<PerformanceAnalytics>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

fetch PoW analytics

fn analytics_create_psuedo_id_if_not_exists<'life0, 'life1, 'async_trait>( &'life0 self, captcha_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Create psuedo ID against campaign ID to publish analytics

fn analytics_get_psuedo_id_from_capmaign_id<'life0, 'life1, 'async_trait>( &'life0 self, captcha_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<String, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Get psuedo ID from campaign ID

fn analytics_get_capmaign_id_from_psuedo_id<'life0, 'life1, 'async_trait>( &'life0 self, psuedo_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<String, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Get campaign ID from psuedo ID

fn analytics_delete_all_records_for_campaign<'life0, 'life1, 'async_trait>( &'life0 self, campaign_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Delete all records for campaign

fn analytics_get_all_psuedo_ids<'life0, 'async_trait>( &'life0 self, page: usize, ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Get all psuedo IDs

fn update_max_nonce_for_level<'life0, 'life1, 'async_trait>( &'life0 self, captcha_key: &'life1 str, difficulty_factor: u32, latest_nonce: u32, ) -> Pin<Box<dyn Future<Output = Result<(), DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Track maximum nonce received against captcha levels

fn get_max_nonce_for_level<'life0, 'life1, 'async_trait>( &'life0 self, captcha_key: &'life1 str, difficulty_factor: u32, ) -> Pin<Box<dyn Future<Output = Result<u32, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Get maximum nonce tracked so far for captcha levels

fn stats_get_num_logs_under_time<'life0, 'async_trait>( &'life0 self, duration: u32, ) -> Pin<Box<dyn Future<Output = Result<usize, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Get number of analytics entries that are under a certain duration

fn stats_get_entry_at_location_for_time_limit_asc<'life0, 'async_trait>( &'life0 self, duration: u32, location: u32, ) -> Pin<Box<dyn Future<Output = Result<Option<usize>, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Get the entry at a location in the list of analytics entires under a certain time limit and sorted in ascending order

Provided Methods

fn analytics_captcha_is_published<'life0, 'life1, 'async_trait>( &'life0 self, campaign_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<bool, DBError>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Get publishing status of pow analytics for captcha ID/ campaign ID